requirements.txt
fastapi==0.105.0
transformer==4.35.2